refactor: rename toggleterm.lua to config/float-term.lua
Move out of plugins/ directory since it no longer depends on any
plugin, and rename to avoid confusion with the toggleterm package.
